requirements.txt
arabic-reshaper==2.1.3
astroid==2.7.2
colorama==0.4.4
dj-database-url==0.5.0
Django==2.2
django-userforeignkey==0.3.0
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.2
platformdirs==2.2.0
psycopg2-binary==2.8.6
pylint==2.10.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.5
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
